ai-welding-server/pyproject.toml
[project]
name = "ai-welding-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.9.0",
    "django==5.2",
    "django-ninja>=1.3.0",
    "dotenv>=0.9.9",
    "openpyxl>=3.1.5",
    "PyJWT>=2.10.1",
    "psycopg2>=2.9.12",
    "pillow>=10.0.0",
]
